[FEATURE] Mutate CSP rules to allow vite-server (#77)
If TYPO3's content security policy is enabled, certain rules need to be loosened
for the Vite dev server to work properly. Also, Vite reads the current `nonce` value
from a meta tag in the html `<head>`, which is now provided as well if the dev server
is active.

This change does not impact the CSP rules on production systems since the special
rules are only applied if the Vite dev server is in use, which is not the case in production.