[FEATURE] Prevent duplicate cache busting with TYPO3 13
By default, TYPO3's asset renderer makes sure that all asset files are served
with an appended timestamp to prevent client-side caching issues (also known
as cache busting). For assets generated by vite, this is really not necessary
because these file names already contain a hash that is derived from the file's
contents and thus serves the same purpose as the timestamp.

TYPO3 13 introduced a new flag for the asset collector API which allows to
disable the default cache busting. This change makes use of that new feature,
while leaving the behavior in TYPO3 12 as-is. With TYPO3 13, this also
(marginally) reduces PHP processing of theses files since TYPO3 doesn't
need to fetch and append the timestamp anymore.

Because of cross-version compatibility, the functional tests unfortunately
have a side effect now: They behave differently depending on the TYPO3
version they are executed in. However, this also tests the real behavior of the
extension, so all in all it serves its purpose.

Resolves: #76