[TASK] Use more coherent ViewHelper names
Since the initial release of the extension, both ViewHelpers had suboptimal
names, which made it unnecessarily complicated for beginners to get started
with vite and TYPO3. This change renames both ViewHelpers (and all
references to them in the documentation). Two class aliases are registered
both on PHP and on DI level, which makes this change non-breaking.

These aliases will probably be kept even for v2, but still it is recommended
to adapt the new naming since it is less confusing than before.

Resolves: #31