[BUGFIX] Avoid double path preparation in icon provider
`getAssetPathFromManifest()` already returns a valid web path,
so it's not necessary to call `getPublicPath()` again, which would
only prepare the path once more. This also avoids mixing old TYPO3
path APIs and the new resource publishing API to be introduced with
TYPO3 v14.

At a later step, `ViteService` will be adjusted to use the new
API.