[BUGFIX] Use ConsumableNonce instead of ConsumableString
A change for TYPO3 v13.4.20 resolved the inheritence dependency
of ConsumableNonce to ConsumableString, which resulted in this
package not having a reference to a nonce value anymore.

The coresponding core change can be found at:
https://review.typo3.org/q/I2de6d67e2639a5701c3a2acb0345e2c9c9b683de

In any way, this package should have used ConsumableNonce in the
first place before this change already, since it is more specific.

Fixes: #131