[BUGFIX] Apply type hint locally in VH to be compatible with Fluid 5
Fluid 5 adds a PHP type for `$renderingContext` in `AbstractViewHelper`,
which means that ViewHelpers can't "re-declare" this property if they
want to stay compatible with older TYPO3 versions. The type hint to
TYPO3's rendering context, which provided `getRequest()` until TYPO3 12,
is now applied locally and can be removed if compatibility for v12 is
dropped.